requirements.txt
torch
torchvision
onnx
onnxruntime
numpy
Pillow
requests
cerebrium
numpy
